feat(prover): extract reusable prover protocol crate
