+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
   "crates/payload",
   "crates/precompiles",
   "crates/primitives",
+  "crates/prover",
   "crates/rpc",
   "crates/sequencer",
   "crates/spf",
@@ -121,7 +122,6 @@ tempo-transaction-pool = { git = "https://github.com/tempoxyz/tempo", rev = "495
 
 # zones
 tempo-zone-contracts = { path = "crates/contracts", default-features = false }
-tempo-zone-prover-enclave = { path = "bin/prover/enclave" }
 zone-chainspec = { path = "crates/chainspec" }
 zone-evm = { path = "crates/evm" }
 zone-l1 = { path = "crates/l1" }
@@ -130,6 +130,7 @@ zone-p2p = { path = "crates/p2p" }
 zone-payload = { path = "crates/payload" }
 zone-precompiles = { path = "crates/precompiles", default-features = false }
 zone-primitives = { path = "crates/primitives", default-features = false }
+zone-prover = { path = "crates/prover" }
 zone-rpc = { path = "crates/rpc" }
 zone-sequencer = { path = "crates/sequencer" }
 zone-spf = { path = "crates/spf" }
