+++ bin/prover/enclave/src/main.rs
@@ -3,9 +3,14 @@ use std::{io, path::PathBuf, process::ExitCode, sync::Arc};
 use alloy_genesis::Genesis;
 use clap::Parser;
 use tempo_chainspec::TempoChainSpec;
-use tempo_zone_prover_enclave::{DEFAULT_MAX_REQUEST_BYTES, DEFAULT_VSOCK_PORT, TrustedChainSpecs};
 use tracing::error;
 use tracing_subscriber::EnvFilter;
+use zone_chainspec::ZoneChainSpec;
+use zone_prover::{
+    DEFAULT_MAX_REQUEST_BYTES, ErrorCode, PROTOCOL_VERSION, ProverConnection, TrustedChainSpecs,
+    VerifyRequest, VerifyResponse, request_error_response,
+};
+use zone_spf::{SpfConfig, prove_zone_batch};
 
 #[tokio::main]
 async fn main() -> ExitCode {
@@ -33,8 +38,8 @@ async fn main() -> ExitCode {
     about = "AWS Nitro Enclave service for the Tempo Zone SPF"
 )]
 struct Cli {
-    /// AF_VSOCK port on which the enclave accepts verification requests.
-    #[arg(long, env = "SPF_VSOCK_PORT", default_value_t = DEFAULT_VSOCK_PORT)]
+    /// Port on which the service accepts verification requests.
+    #[arg(long, env = "SPF_PORT", default_value_t = 5000)]
     port: u32,
 
     /// Maximum accepted JSON request size in bytes.
@@ -48,15 +53,23 @@ struct Cli {
     /// Directory containing trusted custom Tempo genesis JSON files.
     #[arg(long, env = "SPF_TEMPO_GENESIS", value_name = "DIR")]
     tempo_genesis: Option<PathBuf>,
+
+    /// Listen on TCP instead of AF_VSOCK.
+    #[arg(long)]
+    use_tcp: bool,
 }
 
 impl Cli {
     async fn run(self) -> io::Result<()> {
         let specs = self.load_trusted_chain_specs()?;
 
+        if self.use_tcp {
+            return serve_tcp(self.port, self.max_request_bytes, specs).await;
+        }
+
         #[cfg(target_os = "linux")]
         {
-            linux::serve(self.port, self.max_request_bytes, specs).await
+            linux::serve_vsock(self.port, self.max_request_bytes, specs).await
         }
 
         #[cfg(not(target_os = "linux"))]
@@ -102,20 +115,45 @@ impl Cli {
     }
 }
 
+async fn serve_tcp(port: u32, maximum: usize, specs: TrustedChainSpecs) -> io::Result<()> {
+    use tokio::net::TcpListener;
+    use tracing::info;
+
+    let port = u16::try_from(port).map_err(|_| {
+        io::Error::new(
+            io::ErrorKind::InvalidInput,
+            format!("TCP port {port} exceeds the maximum of {}", u16::MAX),
+        )
+    })?;
+    let listener = TcpListener::bind(("127.0.0.1", port)).await?;
+    info!(
+        port,
+        max_request_bytes = maximum,
+        "SPF TCP service listening"
+    );
+
+    loop {
+        let (connection, _peer) = match listener.accept().await {
+            Ok(connection) => connection,
+            Err(error) => {
+                error!(%error, "failed to accept TCP connection");
+                continue;
+            }
+        };
+        handle_connection(connection, maximum, &specs).await;
+    }
+}
+
 #[cfg(target_os = "linux")]
 mod linux {
-    use std::{io, time::Instant};
+    use std::io;
 
-    use futures::{SinkExt as _, StreamExt as _};
     use tokio_vsock::{VMADDR_CID_ANY, VsockAddr, VsockListener};
-    use tracing::{error, info, warn};
+    use tracing::{error, info};
 
     use super::*;
-    use tempo_zone_prover_enclave::{
-        framed, framing_error_response, process_payload, serialize_response,
-    };
 
-    pub(super) async fn serve(
+    pub(super) async fn serve_vsock(
         port: u32,
         maximum: usize,
         specs: TrustedChainSpecs,
@@ -135,37 +173,218 @@ mod linux {
                     continue;
                 }
             };
-            let mut connection = framed(connection, maximum);
-            let started = Instant::now();
-            let payload = match connection.next().await {
-                Some(Ok(payload)) => payload,
-                Some(Err(error)) => {
-                    warn!(%error, "rejected SPF request frame");
-                    let encoded = serialize_response(&framing_error_response(&error, maximum));
-                    if let Err(error) = connection.send(encoded.into()).await {
-                        warn!(%error, "failed to write frame error response");
-                    }
-                    continue;
-                }
-                None => {
-                    warn!("connection closed before sending an SPF request frame");
-                    continue;
-                }
-            };
-            let request_bytes = payload.len();
-            let response = process_payload(&payload, &specs);
-            let encoded = serialize_response(&response);
-            let response_bytes = encoded.len();
-            if let Err(error) = connection.send(encoded.into()).await {
-                warn!(%error, "failed to write SPF response");
-            } else {
-                info!(
-                    request_bytes,
-                    response_bytes,
-                    elapsed_ms = started.elapsed().as_millis(),
-                    "SPF request complete"
-                );
+            handle_connection(connection, maximum, &specs).await;
+        }
+    }
+}
+
+async fn handle_connection<T>(stream: T, maximum: usize, specs: &TrustedChainSpecs)
+where
+    T: tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin,
+{
+    use std::time::Instant;
+
+    use tracing::{info, warn};
+    let mut connection = ProverConnection::new(stream, maximum);
+    let started = Instant::now();
+    let request: VerifyRequest = match connection.receive().await {
+        Ok(Some(request)) => request,
+        Err(error) => {
+            warn!(%error, "rejected SPF request frame");
+            if let Err(error) = connection.send(&request_error_response(&error)).await {
+                warn!(%error, "failed to write frame error response");
             }
+            return;
+        }
+        Ok(None) => {
+            warn!("connection closed before sending an SPF request frame");
+            return;
+        }
+    };
+    let request_bytes = connection.last_received_bytes().unwrap_or_default();
+    let response = process_request(request, specs);
+    match connection.send(&response).await {
+        Ok(response_bytes) => {
+            info!(
+                request_bytes,
+                response_bytes,
+                elapsed_ms = started.elapsed().as_millis(),
+                "SPF request complete"
+            );
+        }
+        Err(error) => warn!(%error, "failed to write SPF response"),
+    }
+}
+
+fn process_request(request: VerifyRequest, specs: &TrustedChainSpecs) -> VerifyResponse {
+    if request.version != PROTOCOL_VERSION {
+        return VerifyResponse::Error {
+            version: PROTOCOL_VERSION,
+            request_id: Some(request.request_id),
+            code: ErrorCode::UnsupportedVersion,
+            message: format!(
+                "unsupported protocol version {}; expected {PROTOCOL_VERSION}",
+                request.version
+            ),
+        };
+    }
+
+    let Some(tempo_spec) = specs.resolve(request.tempo_chain_id) else {
+        return VerifyResponse::Error {
+            version: PROTOCOL_VERSION,
+            request_id: Some(request.request_id),
+            code: ErrorCode::UnsupportedChain,
+            message: format!("unsupported Tempo chain ID {}", request.tempo_chain_id),
+        };
+    };
+    let config = SpfConfig::new(
+        Arc::new(ZoneChainSpec::from(tempo_spec)),
+        request.witness.public_inputs.portal,
+    );
+
+    match prove_zone_batch(&config, request.witness) {
+        Ok(output) => VerifyResponse::Ok {
+            version: PROTOCOL_VERSION,
+            request_id: request.request_id,
+            output,
+        },
+        Err(error) => VerifyResponse::Error {
+            version: PROTOCOL_VERSION,
+            request_id: Some(request.request_id),
+            code: ErrorCode::VerificationFailed,
+            message: error.to_string(),
+        },
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use alloy_consensus::Header;
+    use alloy_primitives::{Address, B256, Bytes};
+    use reth_trie_common::EMPTY_ROOT_HASH;
+    use tempo_primitives::TempoHeader;
+    use zone_spf::{BatchWitness, PublicInputs, TempoStateWitness, ZoneStateWitness};
+
+    use super::*;
+
+    #[test]
+    fn rejects_unsupported_protocol_version() {
+        let request = VerifyRequest {
+            version: 2,
+            request_id: "version-test".into(),
+            tempo_chain_id: 42_431,
+            witness: empty_witness(),
+        };
+        let response = process_request(request, &TrustedChainSpecs::default());
+
+        assert!(matches!(
+            response,
+            VerifyResponse::Error {
+                request_id: Some(id),
+                code: ErrorCode::UnsupportedVersion,
+                ..
+            } if id == "version-test"
+        ));
+    }
+
+    #[test]
+    fn rejects_unknown_chain_before_execution() {
+        let request = VerifyRequest {
+            version: PROTOCOL_VERSION,
+            request_id: "chain-test".into(),
+            tempo_chain_id: 99,
+            witness: empty_witness(),
+        };
+        let response = process_request(request, &TrustedChainSpecs::default());
+
+        assert!(matches!(
+            response,
+            VerifyResponse::Error {
+                request_id: Some(id),
+                code: ErrorCode::UnsupportedChain,
+                ..
+            } if id == "chain-test"
+        ));
+    }
+
+    #[test]
+    fn reports_spf_verification_errors_without_panicking() {
+        let request = VerifyRequest {
+            version: PROTOCOL_VERSION,
+            request_id: "spf-test".into(),
+            tempo_chain_id: 42_431,
+            witness: empty_witness(),
+        };
+        let response = process_request(request, &TrustedChainSpecs::default());
+
+        assert!(matches!(
+            response,
+            VerifyResponse::Error {
+                request_id: Some(id),
+                code: ErrorCode::VerificationFailed,
+                ..
+            } if id == "spf-test"
+        ));
+    }
+
+    #[test]
+    fn accepts_a_configured_custom_chain() {
+        let chain_id = 31_318;
+        let mut genesis = Genesis::default();
+        genesis.config.chain_id = chain_id;
+        let mut specs = TrustedChainSpecs::default();
+        specs
+            .insert(chain_id, Arc::new(TempoChainSpec::from_genesis(genesis)))
+            .unwrap();
+        let request = VerifyRequest {
+            version: PROTOCOL_VERSION,
+            request_id: "custom-chain-test".into(),
+            tempo_chain_id: chain_id,
+            witness: empty_witness(),
+        };
+
+        let response = process_request(request, &specs);
+
+        assert!(matches!(
+            response,
+            VerifyResponse::Error {
+                request_id: Some(id),
+                code: ErrorCode::VerificationFailed,
+                ..
+            } if id == "custom-chain-test"
+        ));
+    }
+
+    fn empty_witness() -> BatchWitness {
+        let tempo_header = TempoHeader {
+            inner: Header {
+                number: 2,
+                state_root: EMPTY_ROOT_HASH,
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+        BatchWitness {
+            public_inputs: PublicInputs {
+                parent_chain_id: 42_431,
+                zone_id: 1,
+                portal: Address::repeat_byte(0x11),
+                tempo_block_number: 2,
+                anchor_block_number: 2,
+                anchor_block_hash: B256::ZERO,
+                expected_withdrawal_batch_index: 3,
+            },
+            parent_header: TempoHeader::default(),
+            zone_blocks: Vec::new(),
+            zone_state_witness: ZoneStateWitness {
+                node_pool: Vec::new(),
+                bytecodes: Vec::new(),
+            },
+            tempo_state_witness: TempoStateWitness {
+                initial_tempo_header_rlp: Bytes::from(alloy_rlp::encode(tempo_header)),
+                node_pool: Vec::new(),
+            },
+            tempo_ancestry_headers: Vec::new(),
         }
     }
 }
