+++ bin/prover/utils/Cargo.toml
@@ -16,9 +16,9 @@ zone-chainspec.workspace = true
 zone-evm.workspace = true
 zone-precompiles = { workspace = true, features = ["std"] }
 zone-primitives.workspace = true
+zone-prover.workspace = true
 zone-rpc.workspace = true
 zone-spf.workspace = true
-tempo-zone-prover-enclave.workspace = true
 
 tempo-alloy.workspace = true
 tempo-chainspec.workspace = true
