+++ crates/prover/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "tempo-zone-prover-enclave"
-description = "AWS Nitro Enclave service for the Tempo Zone SPF"
+name = "zone-prover"
+description = "Prover protocol components for Tempo Zones"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -11,21 +11,16 @@ publish.workspace = true
 workspace = true
 
 [dependencies]
-alloy-genesis.workspace = true
-zone-chainspec.workspace = true
-zone-prover.workspace = true
 zone-spf.workspace = true
 
 tempo-chainspec.workspace = true
 
-clap = { workspace = true, features = ["env"] }
+futures.workspace = true
+serde.workspace = true
 serde_json.workspace = true
+thiserror.workspace = true
 tokio.workspace = true
-tracing.workspace = true
-tracing-subscriber.workspace = true
-
-[target.'cfg(target_os = "linux")'.dependencies]
-tokio-vsock.workspace = true
+tokio-util = { workspace = true, features = ["codec"] }
 
 [dev-dependencies]
 alloy-consensus.workspace = true
