+++ crates/prover/src/connection.rs
@@ -1,0 +1,257 @@
+use std::io;
+
+use futures::{SinkExt as _, StreamExt as _};
+use serde::{Serialize, de::DeserializeOwned};
+use tokio::io::{AsyncRead, AsyncWrite};
+use tokio_util::codec::{Framed, LengthDelimitedCodec, LengthDelimitedCodecError};
+
+use crate::{ErrorCode, PROTOCOL_VERSION, VerifyResponse};
+
+/// A typed connection using the prover's length-delimited JSON protocol.
+pub struct ProverConnection<T> {
+    inner: Framed<T, LengthDelimitedCodec>,
+    maximum: usize,
+    last_received_bytes: Option<usize>,
+}
+
+impl<IO> ProverConnection<IO>
+where
+    IO: AsyncRead + AsyncWrite + Unpin,
+{
+    /// Wraps an I/O stream with the prover protocol and its maximum message size.
+    pub fn new(io: IO, maximum: usize) -> Self {
+        Self {
+            inner: LengthDelimitedCodec::builder()
+                .max_frame_length(maximum)
+                .new_framed(io),
+            maximum,
+            last_received_bytes: None,
+        }
+    }
+
+    /// Returns the encoded size of the most recently received message.
+    pub fn last_received_bytes(&self) -> Option<usize> {
+        self.last_received_bytes
+    }
+
+    /// Serializes and sends a typed message, returning its encoded size.
+    pub async fn send<T: Serialize>(
+        &mut self,
+        message: &T,
+    ) -> Result<usize, ProverConnectionError> {
+        let payload = serde_json::to_vec(message).map_err(ProverConnectionError::Json)?;
+        let bytes = payload.len();
+        self.inner
+            .send(payload.into())
+            .await
+            .map_err(|error| classify_io_error(error, self.maximum))?;
+        Ok(bytes)
+    }
+
+    /// Receives and deserializes a typed message.
+    pub async fn receive<T: DeserializeOwned>(
+        &mut self,
+    ) -> Result<Option<T>, ProverConnectionError> {
+        self.last_received_bytes = None;
+        let Some(payload) = self.inner.next().await else {
+            return Ok(None);
+        };
+        let payload = payload.map_err(|error| classify_io_error(error, self.maximum))?;
+        self.last_received_bytes = Some(payload.len());
+        serde_json::from_slice(&payload)
+            .map(Some)
+            .map_err(ProverConnectionError::Json)
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum ProverConnectionError {
+    #[error("message exceeds the maximum of {maximum} bytes")]
+    MessageTooLarge { maximum: usize },
+    #[error("message frame is truncated")]
+    TruncatedFrame,
+    #[error("message JSON is invalid: {0}")]
+    Json(#[source] serde_json::Error),
+    #[error("connection I/O failed: {0}")]
+    Io(#[source] io::Error),
+}
+
+fn classify_io_error(error: io::Error, maximum: usize) -> ProverConnectionError {
+    if error
+        .get_ref()
+        .is_some_and(|source| source.is::<LengthDelimitedCodecError>())
+    {
+        ProverConnectionError::MessageTooLarge { maximum }
+    } else if error.kind() == io::ErrorKind::Other
+        && error.to_string() == "bytes remaining on stream"
+    {
+        ProverConnectionError::TruncatedFrame
+    } else {
+        ProverConnectionError::Io(error)
+    }
+}
+
+/// Converts a request receive error into a prover protocol response.
+pub fn request_error_response(error: &ProverConnectionError) -> VerifyResponse {
+    let (code, message) = match error {
+        ProverConnectionError::MessageTooLarge { maximum } => (
+            ErrorCode::RequestTooLarge,
+            format!("request payload exceeds the maximum of {maximum} bytes"),
+        ),
+        ProverConnectionError::TruncatedFrame => (
+            ErrorCode::TruncatedFrame,
+            "request frame is truncated".into(),
+        ),
+        ProverConnectionError::Json(error) => (
+            ErrorCode::MalformedRequest,
+            format!("invalid request JSON: {error}"),
+        ),
+        ProverConnectionError::Io(error) => (
+            ErrorCode::InternalError,
+            format!("frame I/O failed: {error}"),
+        ),
+    };
+    VerifyResponse::Error {
+        version: PROTOCOL_VERSION,
+        request_id: None,
+        code,
+        message,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use alloy_consensus::Header;
+    use alloy_primitives::{Address, B256, Bytes};
+    use reth_trie_common::EMPTY_ROOT_HASH;
+    use tempo_primitives::TempoHeader;
+    use tokio::io::AsyncWriteExt as _;
+    use zone_spf::{BatchWitness, PublicInputs, TempoStateWitness, ZoneStateWitness};
+
+    use super::*;
+    use crate::VerifyRequest;
+
+    #[tokio::test]
+    async fn request_round_trip() {
+        let maximum = 1024 * 1024;
+        let (client, server) = tokio::io::duplex(maximum);
+        let mut client = ProverConnection::new(client, maximum);
+        let mut server = ProverConnection::new(server, maximum);
+        let request = VerifyRequest {
+            version: PROTOCOL_VERSION,
+            request_id: "round-trip".into(),
+            tempo_chain_id: 42_431,
+            witness: empty_witness(),
+        };
+        let sent_bytes = client.send(&request).await.unwrap();
+        let received: VerifyRequest = server.receive().await.unwrap().unwrap();
+
+        assert_eq!(received.version, PROTOCOL_VERSION);
+        assert_eq!(received.request_id, "round-trip");
+        assert_eq!(received.tempo_chain_id, 42_431);
+        assert_eq!(server.last_received_bytes(), Some(sent_bytes));
+
+        let response = VerifyResponse::Error {
+            version: PROTOCOL_VERSION,
+            request_id: Some(received.request_id),
+            code: ErrorCode::VerificationFailed,
+            message: "round-trip".into(),
+        };
+        let sent_bytes = server.send(&response).await.unwrap();
+        let received: VerifyResponse = client.receive().await.unwrap().unwrap();
+        assert!(matches!(
+            received,
+            VerifyResponse::Error {
+                request_id: Some(id),
+                code: ErrorCode::VerificationFailed,
+                ..
+            } if id == "round-trip"
+        ));
+        assert_eq!(client.last_received_bytes(), Some(sent_bytes));
+    }
+
+    #[tokio::test]
+    async fn rejects_oversized_and_truncated_frames() {
+        let (mut writer, reader) = tokio::io::duplex(1024);
+        writer.write_all(&10_u32.to_be_bytes()).await.unwrap();
+        let error = ProverConnection::new(reader, 4)
+            .receive::<VerifyRequest>()
+            .await
+            .unwrap_err();
+        assert!(matches!(
+            request_error_response(&error),
+            VerifyResponse::Error {
+                code: ErrorCode::RequestTooLarge,
+                ..
+            }
+        ));
+
+        let (mut writer, reader) = tokio::io::duplex(1024);
+        writer.write_all(&[0, 0, 0, 2, 1]).await.unwrap();
+        writer.shutdown().await.unwrap();
+        let error = ProverConnection::new(reader, 4)
+            .receive::<VerifyRequest>()
+            .await
+            .unwrap_err();
+        assert!(matches!(
+            request_error_response(&error),
+            VerifyResponse::Error {
+                code: ErrorCode::TruncatedFrame,
+                ..
+            }
+        ));
+    }
+
+    #[tokio::test]
+    async fn malformed_json_has_a_stable_error() {
+        let (mut writer, reader) = tokio::io::duplex(1024);
+        writer.write_all(&8_u32.to_be_bytes()).await.unwrap();
+        writer.write_all(b"not json").await.unwrap();
+        let error = ProverConnection::new(reader, 1024)
+            .receive::<VerifyRequest>()
+            .await
+            .unwrap_err();
+        let response = request_error_response(&error);
+        assert!(matches!(
+            response,
+            VerifyResponse::Error {
+                request_id: None,
+                code: ErrorCode::MalformedRequest,
+                ..
+            }
+        ));
+    }
+
+    fn empty_witness() -> BatchWitness {
+        let tempo_header = TempoHeader {
+            inner: Header {
+                number: 2,
+                state_root: EMPTY_ROOT_HASH,
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+        BatchWitness {
+            public_inputs: PublicInputs {
+                parent_chain_id: 42_431,
+                zone_id: 1,
+                portal: Address::repeat_byte(0x11),
+                tempo_block_number: 2,
+                anchor_block_number: 2,
+                anchor_block_hash: B256::ZERO,
+                expected_withdrawal_batch_index: 3,
+            },
+            parent_header: TempoHeader::default(),
+            zone_blocks: Vec::new(),
+            zone_state_witness: ZoneStateWitness {
+                node_pool: Vec::new(),
+                bytecodes: Vec::new(),
+            },
+            tempo_state_witness: TempoStateWitness {
+                initial_tempo_header_rlp: Bytes::from(alloy_rlp::encode(tempo_header)),
+                node_pool: Vec::new(),
+            },
+            tempo_ancestry_headers: Vec::new(),
+        }
+    }
+}
